TRUNK-6332: Hibernate 6.x and spring 6.x upgrade (#5062)
* TRUNK-6332: Hibernate 6.x and spring 6.x upgrade

* TRUNK-6332: Upgrade h2 to 2.x

* TRUNK-6332: Fix obs test

* TRUNK-6332: Fix obs test

* TRUNK-6332: Remove java 8 from pipelines

* TRUNK-6332: Migrate ProviderRole to jakarta

* TRUNK-6332: Migrate StartupFilterTest to jakarta

* TRUNK-6332: Remove wildcard imports

* TRUNK-6332: getFormField_shouldNotFailWithNullIgnoreFormFieldsArgument test

* TRUNK-6332: Fix cargo, migrate more javax and remove tomcat7

* TRUNK-6332: Fix encounter join issue when ordering by visit fields

* TRUNK-6332: Update jasper version after merge

+++ api/src/main/java/org/openmrs/api/impl/AdministrationServiceImpl.java
@@ -876,7 +876,7 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	 */
 	@Override
 	@Transactional(readOnly = true)
-	public int getMaximumPropertyLength(Class<? extends OpenmrsObject> aClass, String fieldName) {
+	public long getMaximumPropertyLength(Class<? extends OpenmrsObject> aClass, String fieldName) {
 		return dao.getMaximumPropertyLength(aClass, fieldName);
 	}
